requirements.txt
Package                        Version
------------------------------ ---------
anyio                          3.4.0
appnope                        0.1.2
argon2-cffi                    21.1.0
attrs                          21.2.0
Babel                          2.9.1
backcall                       0.2.0
bleach                         4.1.0
cachetools                     4.2.4
certifi                        2021.10.8
cffi                           1.15.0
charset-normalizer             2.0.7
click                          8.0.3
cloudevents                    1.2.0
debugpy                        1.5.1
decorator                      5.1.0
defusedxml                     0.7.1
deprecation                    2.1.0
entrypoints                    0.3
Flask                          2.0.2
functions-framework            3.0.0
google-api-core                2.3.0
google-auth                    2.3.3
google-cloud-appengine-logging 1.1.0
google-cloud-audit-log         0.2.0
google-cloud-bigquery          2.31.0
google-cloud-core              2.2.1
google-cloud-firestore         2.3.4
google-cloud-logging           2.7.0
google-crc32c                  1.3.0
google-resumable-media         2.1.0
googleapis-common-protos       1.54.0
grpc-google-iam-v1             0.12.3
grpcio                         1.42.0
grpcio-status                  1.42.0
gunicorn                       20.1.0
idna                           3.3
ipykernel                      6.5.1
ipython                        7.29.0
ipython-genutils               0.2.0
ipywidgets                     7.6.5
itsdangerous                   2.0.1
jedi                           0.18.1
Jinja2                         3.0.3
json5                          0.9.6
jsonschema                     4.2.1
jupyter                        1.0.0
jupyter-client                 7.1.0
jupyter-console                6.4.0
jupyter-core                   4.9.1
jupyter-server                 1.12.0
jupyterlab                     3.2.4
jupyterlab-pygments            0.1.2
jupyterlab-server              2.8.2
jupyterlab-widgets             1.0.2
MarkupSafe                     2.0.1
matplotlib-inline              0.1.3
mistune                        0.8.4
nbclassic                      0.3.4
nbclient                       0.5.9
nbconvert                      6.3.0
nbformat                       5.1.3
nest-asyncio                   1.5.1
notebook                       6.4.6
numpy                          1.21.4
oauthlib                       3.1.1
packaging                      21.3
pandas                         1.3.4
pandocfilters                  1.5.0
parso                          0.8.2
pexpect                        4.8.0
pickleshare                    0.7.5
pip                            21.3.1
prometheus-client              0.12.0
prompt-toolkit                 3.0.22
proto-plus                     1.19.8
protobuf                       3.19.1
ptyprocess                     0.7.0
pyarrow                        6.0.1
pyasn1                         0.4.8
pyasn1-modules                 0.2.8
pycparser                      2.21
Pygments                       2.10.0
pyparsing                      3.0.6
pyrsistent                     0.18.0
python-dateutil                2.8.2
pytz                           2021.3
pyzmq                          22.3.0
qtconsole                      5.2.1
QtPy                           1.11.2
requests                       2.26.0
requests-oauthlib              1.3.0
rsa                            4.8
Send2Trash                     1.8.0
setuptools                     58.0.4
six                            1.16.0
sniffio                        1.2.0
terminado                      0.12.1
testpath                       0.5.0
tornado                        6.1
traitlets                      5.1.1
tweepy                         4.4.0
urllib3                        1.26.7
watchdog                       1.0.2
wcwidth                        0.2.5
webencodings                   0.5.1
websocket-client               1.2.1
Werkzeug                       2.0.2
wheel                          0.37.0
widgetsnbextension             3.5.2
